requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.8
django-redis==4.4.3
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gevent==1.4.0
greenlet==0.4.15
grequests==0.4.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2019.3
redis==2.10.6
requests==2.22.0
requests-futures==1.0.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
